move mapping to own file
