+++ src/background/api/mapping.js
@@ -1,3 +1,61 @@
+import { quickStorage } from '../settings/quickStorage';
+
+export function mapDataToInternalFormat(data) {
+	const { viewer, ...repos } = data;
+
+	// Loop through each reeturned repo
+	return Object.values(repos).map((repo) => {
+		const oldRepo = quickStorage.repositories.find(
+			(oldRepo) => oldRepo.url === repo.url
+		);
+
+		// TODO: Can we do the transfer of data inside this function?
+		// instead of doing seperate loop
+
+		// TODO: Hvis ikke vi har et gammelt repo eller et gammelt
+		// tall så har vi nok lagt til et nytt repo
+		// Bør vi sete ting til lest? Blir vel feil å starte opp
+		// med å sette alt til ulest?
+		const oldTotalItemsNumber = oldRepo?.totalItemsNumber;
+		// const newTotalItemsNumber = calculateMaxNumber(repo);
+		// console.log(quickStorage.repositories);
+		// console.log(
+		// 	'🚀 => Object.keys => newTotalItemsNumber',
+		// 	newTotalItemsNumber
+		// );
+		// console.log(
+		// 	'🚀 => Object.keys => oldTotalItemsNumber',
+		// 	oldTotalItemsNumber
+		// );
+
+		// TODO: NaN issue med nora setup av en eller annen grunn?
+		// KAn være fordi det ikke eksister noen issues der fra før
+		// Men vi får jo for pulls, så er litt rart
+
+		// Mapping data from items
+		const issues = listItems(repo.issues, viewer, oldTotalItemsNumber);
+		const pullRequests = listItems(
+			repo.pullRequests,
+			viewer,
+			oldTotalItemsNumber
+		);
+
+		return {
+			name: repo.name,
+			owner: repo.owner.login,
+			url: repo.url,
+			collapsed: true,
+			// totalItemsNumber: newTotalItemsNumber,
+			totalItems: {
+				issues: repo.issues.totalCount,
+				pullRequests: repo.pullRequests.totalCount,
+			},
+			issues,
+			pullRequests,
+		};
+	});
+}
+
 export function listItems(element, { login }, totalItemNumber) {
 	return element.edges.map(({ node: item }) => {
 		return {
